Fix: Round 5 audit — 1 critical + 2 medium + 3 low issues
Critical:
- channel.c: uint16_t truncation of hash_size → uint32_t types.h field,
  cap 65536→65535, remove (uint16_t) cast

Medium:
- channel.c: last_global_heartbeat set only on PONG, not PING
- main.c: add !mac_is_zero() guard to peer MAC auto-learning

Low:
- channel.c: ignore duplicate SYN on ESTABLISHED (no SYN_RCVD regression)
- main.c: clarify heartbeat params read from config at check time
- main.c: remove channel_count>0 guard on KCP drain loop during cleanup

+++ src/types.h
@@ -297,7 +297,7 @@ typedef struct {
 
     /* 通道管理 */
     channel_t     **channel_hash;                /* 通道哈希表（动态分配） */
-    uint16_t        channel_hash_size;           /* 哈希表桶数 */
+    uint32_t        channel_hash_size;           /* 哈希表桶数 */
     int             channel_count;               /* 当前活跃通道数 */
 
     /* epoll */
